# sample/Demo.Api.Generated/Endpoints/AddressesController.cs
using System;
using System.CodeDom.Compiler;
using System.Collections.Generic;
using System.Threading;
using System.Threading.Tasks;
using Demo.Api.Generated.Contracts.Addresses;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;

//------------------------------------------------------------------------------
// This code was auto-generated by ApiGenerator 1.0.0.0.
//
// Changes to this file may cause incorrect behavior and will be lost if
// the code is regenerated.
//------------------------------------------------------------------------------
//
// ReSharper disable once CheckNamespace
namespace Demo.Api.Generated.Endpoints
{
    /// <summary>
    /// Endpoint definitions.
    /// Area: Addresses.
    /// </summary>
    [ApiController]
    [Route("api/v1/addresses")]
    [GeneratedCode("ApiGenerator", "1.0.0.0")]
    public class AddressesController : ControllerBase
    {
        /// <summary>
        /// Description: Get addresses by postal code.
        /// Operation: GetAddressesByPostalCodes.
        /// Area: Addresses.
        /// </summary>
        [HttpGet("{postalCode}")]
        [ProducesResponseType(typeof(List<Contracts.Address>), StatusCodes.Status200OK)]
        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
        public Task<ActionResult> GetAddressesByPostalCodesAsync([FromServices] IGetAddressesByPostalCodesHandler handler, CancellationToken cancellationToken)
        {
            if (handler is null)
            {
                throw new ArgumentNullException(nameof(handler));
            }

            return InvokeGetAddressesByPostalCodesAsync(handler, cancellationToken);
        }

        private static async Task<ActionResult> InvokeGetAddressesByPostalCodesAsync([FromServices] IGetAddressesByPostalCodesHandler handler, CancellationToken cancellationToken)
        {
            return await handler.ExecuteAsync(cancellationToken);
        }
    }
}